Create a `CreateKeywordResult` service
